# perf: Shuffle followed by Take/Skip/TakeLast/SkipLast