--- a/src/ZLinq/Linq/Shuffle.SkipShuffle.cs
+++ b/src/ZLinq/Linq/Shuffle.SkipShuffle.cs
@@ -120,6 +120,7 @@ namespace ZLinq.Linq
         }
 
         [MemberNotNull(nameof(buffer))]
+        [MethodImpl(MethodImplOptions.NoInlining)]
         void InitBuffer()
         {
             if (buffer == null)