Similar optimization to order and take/skip combo.


## Considerations

`shuffle().take()` will compacting enumerator connection from:
- array -> shuffle -> take

to:
- array -> skipShuffle

intermediate shuffle will be hidden orphan branch, but it will never take internal buffer until consumed so it won't be a problem.


## Benchmarks

Add .NET 10.0 as a target and the following code at the beginning of Main().
```cs
BenchmarkRunner.Run(typeof(ShuffleBench));
return 0;
```

Then, set build target to `net10.0` and go ahead.

| Method          | N      | M      | Mean         | Error        | StdDev       |
|---------------- |------- |------- |-------------:|-------------:|-------------:|
| ShuffleTakeLinq | 1000   | 10     |   3,560.3 ns |     16.58 ns |     14.69 ns |
| ShuffleTakeLinq | 1000   | 100    |   5,853.1 ns |     31.23 ns |     29.22 ns |
| ShuffleTakeLinq | 1000   | 1000   |   6,365.5 ns |     79.05 ns |     70.08 ns |
| ShuffleTakeLinq | 1000   | 100000 |   6,278.3 ns |     28.74 ns |     25.48 ns |
| ShuffleTakeZL   | 1000   | 10     |     106.1 ns |      0.61 ns |      0.57 ns |
| ShuffleTakeZL   | 1000   | 100    |     551.5 ns |      1.41 ns |      1.32 ns |
| ShuffleTakeZL   | 1000   | 1000   |   5,197.1 ns |     28.21 ns |     23.56 ns |
| ShuffleTakeZL   | 1000   | 100000 |   5,442.6 ns |     23.21 ns |     21.71 ns |
|                 |        |        |              |              |              |
| ShuffleTakeLinq | 100000 | 10     | 303,312.2 ns |  1,473.92 ns |  1,378.71 ns |
| ShuffleTakeLinq | 100000 | 100    | 310,066.7 ns |  1,528.47 ns |  1,276.34 ns |
| ShuffleTakeLinq | 100000 | 1000   | 351,249.2 ns |  3,069.96 ns |  2,721.44 ns |
| ShuffleTakeLinq | 100000 | 100000 | 790,944.2 ns | 13,770.01 ns | 13,523.99 ns |
| ShuffleTakeZL   | 100000 | 10     |  11,649.3 ns |    232.66 ns |    228.51 ns |
| ShuffleTakeZL   | 100000 | 100    |  12,209.2 ns |    238.38 ns |    222.98 ns |
| ShuffleTakeZL   | 100000 | 1000   |  17,677.4 ns |    322.27 ns |    285.69 ns |
| ShuffleTakeZL   | 100000 | 100000 | 493,672.9 ns |  2,925.18 ns |  2,442.66 ns |